+++ dspic33ck-curiosity-crc-advanced.X/print.h
@@ -52,7 +52,7 @@ enum
 
 #define timerCountToMs(timerCount, timerPeriod) ((timerCount / (double)Timer1.PeriodGet()) * (double)timerPeriod)
 
-#define calculateInstructions(softwareCRCCalculationTime) (CLOCK_InstructionFrequencyGet() * softwareCRCCalculationTime)
+#define calculateInstructions(softwareCRCCalculationTime) ((double)CLOCK_InstructionFrequencyGet() * softwareCRCCalculationTime)
 
 //Section: Printing APIs
